requirements.txt
# Core dependencies
cognee>=0.1.0
cognee-community-hybrid-adapter-falkor>=0.1.0
mcp>=0.9.0
python-dotenv>=1.0.0

# LLM providers
openai>=1.0.0
google-generativeai>=0.3.0

# HTTP client
httpx>=0.25.0

# Async utilities
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0

# Data handling
pydantic>=2.0.0
pydantic-settings>=2.0.0
